feat(DTFS2-7578): refactor helpers into new folder
