+++ portal-api/src/v1/gef/services/clone/helpers/generate-cloned-exporter.ts
@@ -1,12 +1,12 @@
 import { AnyObject } from '@ukef/dtfs2-common';
-import { exporterStatus } from '../../controllers/validation/exporter';
+import { exporterStatus } from '../../../controllers/validation/exporter';
 
 /**
  * Clones the exporter
  * @param currentExporter the current exporter
  * @returns the exporter with updatedAt set to now
  */
-export const cloneExporter = (currentExporter: AnyObject) => {
+export const generateClonedExporter = (currentExporter: AnyObject) => {
   const clonedExporter = currentExporter;
 
   clonedExporter.updatedAt = Date.now();
